+++ source/systems/spritegraphics.d
@@ -70,7 +70,7 @@ class SpriteGraphics : System!Sprite
 
   override void updateFromEntities()
   {
-    foreach (ulong index, Entity entity; entityForIndex)
+    foreach (uint index, Entity entity; entityForIndex)
     {
       //currentStates[index].velocity = entity.vectors["velocity"];
       components[index].position = vec2(entity.values["position"].to!(float[2]));
